requirements.txt
-i http://pypi.doubanio.com/simple/
Django~=1.11
mistune==0.8.4
xadmin==0.6.1
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
pillow==5.1.0
djangorestframework==3.8.2
coreapi==2.3.3

